pyproject.toml
[project]
name = "quantum-docking"
version = "0.0.1"
authors = [{name="Mathieu Garrigues", email="mathieu.garrigues@pasqal.com"}, {name="Victor Onofre", email="victor.onofre@pasqal.com"},]
description = "As part of the Blaise Pascal's [Re]Generative Quantum Challenge, we attempt to use neutral atoms to speed up molecular docking."
readme = "README.md"
requires-python = ">=3.10"
classifiers = ["Programming Language :: Python :: 3",
    "License :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",]

[project.urls]
"Homepage" = "https://pasqal-quantum-challenge-platform.bemyapp.com/#/projects/6508735070156e9993122511"

#================================================================

[tool.black]
line-length = 100

[tool.pycln]
all = true

[tool.isort]
ensure_newline_before_comments = true
force_grid_wrap = 0
include_trailing_comma = true
line_length = 100
multi_line_output = 3
use_parentheses = true

[tool.ruff]
line-length = 100

[tool.poetry]
name = "quantum-docking"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
jupyter = "^1.0.0"
networkx = "2.8.8"
scipy = "<1.13"
pulser = "^0.19.0"
rdkit = "^2025.3.3"
pmapper = "^1.1.1"
lxml = "^5.3.0"
ipympl = "^0.9.7"
py3dmol = "^2.5.0"
mordred = "^1.2.0"
tk = "^0.1.0"
pyvista = "^0.45.2"
biopython = "^1.85"
seaborn = "^0.13.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
